Data.Discrimination

Type

newtype Disc a

Discriminator

Constructors

Disc

Fields

runDisc :: forall b. [(a, b)] -> [[b]]

desc :: Disc a -> Disc a

Discriminators

Unordered

class Grouping a where

Eq equipped with a compatible stable unordered discriminator.

Minimal complete definition

Nothing

Methods

grouping :: Disc a

class Grouping1 f where

Minimal complete definition

Nothing

Methods

grouping1 :: Disc a -> Disc (f a)

groupingBag :: Foldable f => Disc k -> Disc (f k)

Construct an stable unordered discriminator that partitions into equivalence classes based on the equivalence of keys as a multiset.

groupingSet :: Foldable f => Disc k -> Disc (f k)

Construct an stable unordered discriminator that partitions into equivalence classes based on the equivalence of keys as a set.

groupingEq :: Grouping a => a -> a -> Bool

Valid definition for (==) in terms of Grouping.

Ordered

class Grouping a => Sorting a where

Ord equipped with a compatible stable, ordered discriminator.

Minimal complete definition

Nothing

Methods

sorting :: Disc a

class Grouping1 f => Sorting1 f where

Minimal complete definition

Nothing

Methods

sorting1 :: Disc a -> Disc (f a)

sortingBag :: Foldable f => Disc k -> Disc (f k)

Construct a stable ordered discriminator that sorts a list as multisets of elements from another stable ordered discriminator.

The resulting discriminator only cares about the set of keys and their multiplicity, and is sorted as if we'd sorted each key in turn before comparing.

sortingSet :: Foldable f => Disc k -> Disc (f k)

Construct a stable ordered discriminator that sorts a list as sets of elements from another stable ordered discriminator.

The resulting discriminator only cares about the set of keys, and is sorted as if we'd sorted each key in turn before comparing.

sortingCompare :: Sorting a => a -> a -> Ordering

Valid definition for compare in terms of Sorting.

Combinators

Unordered

nub :: Grouping a => [a] -> [a]

O(n). This upgrades nub from Data.List from O(n^2) to O(n) by using unordered discrimination.

nub = nubWith id
nub as = head <$> group as

nubWith :: Grouping b => (a -> b) -> [a] -> [a]

O(n). nub with a Schwartzian transform.

nubWith f as = head <$> groupWith f as

group :: Grouping a => [a] -> [[a]]

O(n). Similar to group, except we do not require groups to be clustered.

This combinator still operates in linear time, at the expense of productivity.

The result equivalence classes are _not_ sorted, but the grouping is stable.

group = groupWith id

groupWith :: Grouping b => (a -> b) -> [a] -> [[a]]

O(n). This is a replacement for groupWith using discrimination.

The result equivalence classes are _not_ sorted, but the grouping is stable.

Ordered

sort :: Sorting a => [a] -> [a]

O(n). Sort a list using discrimination.

sort = sortWith id

sortWith :: Sorting b => (a -> b) -> [a] -> [a]

O(n). Sort a list with a Schwartzian transformation by using discrimination.

This linear time replacement for sortWith and sortOn uses discrimination.

Containers

toMap :: Sorting k => [(k, v)] -> Map k v

O(n). Construct a Map.

This is an asymptotically faster version of fromList, which exploits ordered discrimination.

>>> toMap [] == empty
True

>>> toMap [(5,"a"), (3 :: Int,"b"), (5, "c")]
fromList [(5,"c"), (3,"b")]

>>> toMap [(5,"c"), (3,"b"), (5 :: Int, "a")]
fromList [(5,"a"), (3,"b")]

toMapWith :: Sorting k => (v -> v -> v) -> [(k, v)] -> Map k v

O(n). Construct a Map, combining values.

This is an asymptotically faster version of fromListWith, which exploits ordered discrimination.

(Note: values combine in anti-stable order for compatibility with fromListWith)

>>> toMapWith (++) [(5,"a"), (5,"b"), (3,"b"), (3,"a"), (5 :: Int,"c")]
fromList [(3, "ab"), (5, "cba")]

>>> toMapWith (++) [] == empty
True

toMapWithKey :: Sorting k => (k -> v -> v -> v) -> [(k, v)] -> Map k v

O(n). Construct a Map, combining values with access to the key.

This is an asymptotically faster version of fromListWithKey, which exploits ordered discrimination.

(Note: the values combine in anti-stable order for compatibility with fromListWithKey)

>>> let f key new_value old_value = show key ++ ":" ++ new_value ++ "|" ++ old_value
>>> toMapWithKey f [(5,"a"), (5,"b"), (3,"b"), (3,"a"), (5 :: Int,"c")]
fromList [(3, "3:a|b"), (5, "5:c|5:b|a")]

>>> toMapWithKey f [] == empty
True

toIntMap :: [(Int, v)] -> IntMap v

O(n). Construct an IntMap.

>>> toIntMap [] == empty
True

>>> toIntMap [(5,"a"), (3,"b"), (5, "c")]
fromList [(5,"c"), (3,"b")]

>>> toIntMap [(5,"c"), (3,"b"), (5, "a")]
fromList [(5,"a"), (3,"b")]

toIntMapWith :: (v -> v -> v) -> [(Int, v)] -> IntMap v

O(n). Construct an IntMap, combining values.

This is an asymptotically faster version of fromListWith, which exploits ordered discrimination.

(Note: values combine in anti-stable order for compatibility with fromListWith)

>>> toIntMapWith (++) [(5,"a"), (5,"b"), (3,"b"), (3,"a"), (5,"c")]
fromList [(3, "ab"), (5, "cba")]

>>> toIntMapWith (++) [] == empty
True

toIntMapWithKey :: (Int -> v -> v -> v) -> [(Int, v)] -> IntMap v

O(n). Construct a Map, combining values with access to the key.

This is an asymptotically faster version of fromListWithKey, which exploits ordered discrimination.

(Note: the values combine in anti-stable order for compatibility with fromListWithKey)

>>> let f key new_value old_value = show key ++ ":" ++ new_value ++ "|" ++ old_value
>>> toIntMapWithKey f [(5,"a"), (5,"b"), (3,"b"), (3,"a"), (5,"c")]
fromList [(3, "3:a|b"), (5, "5:c|5:b|a")]

>>> toIntMapWithKey f [] == empty
True

toSet :: Sorting k => [k] -> Set k

O(n). Construct a Set in linear time.

This is an asymptotically faster version of fromList, which exploits ordered discrimination.

toIntSet :: [Int] -> IntSet

O(n). Construct an IntSet in linear time.

This is an asymptotically faster version of fromList, which exploits ordered discrimination.

Joins

joining

Arguments

:: Disc d
the discriminator to use

-> ([a] -> [b] -> c)
how to join two tables

-> (a -> d)
selector for the left table

-> (b -> d)
selector for the right table

-> [a]
left table

-> [b]
right table

-> [c]

O(n). Perform a full outer join while explicit merging of the two result tables a table at a time.

The results are grouped by the discriminator.

inner

Arguments

:: Disc d
the discriminator to use

-> (a -> b -> c)
how to join two rows

-> (a -> d)
selector for the left table

-> (b -> d)
selector for the right table

-> [a]
left table

-> [b]
right table

-> [[c]]

O(n). Perform an inner join, with operations defined one row at a time.

The results are grouped by the discriminator.

This takes operation time linear in both the input and result sets.

outer

Arguments

:: Disc d
the discriminator to use

-> (a -> b -> c)
how to join two rows

-> (a -> c)
row present on the left, missing on the right

-> (b -> c)
row present on the right, missing on the left

-> (a -> d)
selector for the left table

-> (b -> d)
selector for the right table

-> [a]
left table

-> [b]
right table

-> [[c]]

O(n). Perform a full outer join with operations defined one row at a time.

The results are grouped by the discriminator.

This takes operation time linear in both the input and result sets.

leftOuter

Arguments

:: Disc d
the discriminator to use

-> (a -> b -> c)
how to join two rows

-> (a -> c)
row present on the left, missing on the right

-> (a -> d)
selector for the left table

-> (b -> d)
selector for the right table

-> [a]
left table

-> [b]
right table

-> [[c]]

O(n). Perform a left outer join with operations defined one row at a time.

The results are grouped by the discriminator.

This takes operation time linear in both the input and result sets.

rightOuter

Arguments

:: Disc d
the discriminator to use

-> (a -> b -> c)
how to join two rows

-> (b -> c)
row present on the right, missing on the left

-> (a -> d)
selector for the left table

-> (b -> d)
selector for the right table

-> [a]
left table

-> [b]
right table

-> [[c]]

O(n). Perform a right outer join with operations defined one row at a time.

The results are grouped by the discriminator.

This takes operation time linear in both the input and result sets.