Data.Discrimination.Grouping

Documentation

newtype Group a

Productive Stable Unordered Discriminator

Constructors

Group

Fields

getGroup :: forall m b. PrimMonad m => (b -> m (b -> m ())) -> m (a -> b -> m ())

class Grouping a where

Eq equipped with a compatible stable unordered discriminator.

Minimal complete definition

Nothing

Methods

grouping :: Group a

For every surjection f,

contramap f grouping ≡ grouping

class Grouping1 f where

Minimal complete definition

Nothing

Methods

grouping1 :: Group a -> Group (f a)

Combinators

nub :: Grouping a => [a] -> [a]

O(n). This upgrades nub from Data.List from O(n^2) to O(n) by using productive unordered discrimination.

nub = nubWith id
nub as = head <$> group as

nubWith :: Grouping b => (a -> b) -> [a] -> [a]

O(n). Online nub with a Schwartzian transform.

nubWith f as = head <$> groupWith f as

group :: Grouping a => [a] -> [[a]]

O(n). Similar to group, except we do not require groups to be clustered.

This combinator still operates in linear time, at the expense of storing history.

The result equivalence classes are _not_ sorted, but the grouping is stable.

group = groupWith id

groupWith :: Grouping b => (a -> b) -> [a] -> [[a]]

O(n). This is a replacement for groupWith using discrimination.

The result equivalence classes are _not_ sorted, but the grouping is stable.

groupingEq :: Grouping a => a -> a -> Bool

Valid definition for (==) in terms of Grouping.

runGroup :: Group a -> [(a, b)] -> [[b]]

Internals

groupingNat :: Int -> Group Int