Data.Discrimination

Discrimination

class Decidable f => Discriminating f where

Methods

disc :: f a -> [(a, b)] -> [[b]]

Unordered

newtype Group a

Productive Stable Unordered Discriminator

Constructors

Group

Fields

getGroup :: forall m b. PrimMonad m => (b -> m (b -> m ())) -> m (a -> b -> m ())

class Grouping a where

Eq equipped with a compatible stable unordered discriminator.

Law:

groupingEq x y ≡ (x == y)

Note: Eq is a moral super class of Grouping. It isn't because of some missing instances.

Minimal complete definition

Nothing

Methods

grouping :: Group a

For every surjection f,

contramap f grouping ≡ grouping

default grouping :: Deciding Grouping a => Group a

class Grouping1 f where

Minimal complete definition

Nothing

Methods

grouping1 :: Group a -> Group (f a)

default grouping1 :: Deciding1 Grouping f => Group a -> Group (f a)

nub :: Grouping a => [a] -> [a]

O(n). This upgrades nub from Data.List from O(n^2) to O(n) by using productive unordered discrimination.

nub = nubWith id
nub as = head <$> group as

nubWith :: Grouping b => (a -> b) -> [a] -> [a]

O(n). Online nub with a Schwartzian transform.

nubWith f as = head <$> groupWith f as

group :: Grouping a => [a] -> [[a]]

O(n). Similar to group, except we do not require groups to be clustered.

This combinator still operates in linear time, at the expense of storing history.

The result equivalence classes are not sorted, but the grouping is stable.

group = groupWith id

groupWith :: Grouping b => (a -> b) -> [a] -> [[a]]

O(n). This is a replacement for groupWith using discrimination.

The result equivalence classes are not sorted, but the grouping is stable.

runGroup :: Group a -> [(a, b)] -> [[b]]

groupingEq :: Grouping a => a -> a -> Bool

Valid definition for (==) in terms of Grouping.

Ordered

newtype Sort a

Stable Ordered Discriminator

Constructors

Sort

Fields

runSort :: forall b. [(a, b)] -> [[b]]

class Grouping a => Sorting a where

Ord equipped with a compatible stable, ordered discriminator.

Law:

sortingCompare x y ≡ compare x y

Minimal complete definition

Nothing

Methods

sorting :: Sort a

For every strictly monotone-increasing function f:

contramap f sorting ≡ sorting

default sorting :: Deciding Sorting a => Sort a

class Grouping1 f => Sorting1 f where

Minimal complete definition

Nothing

Methods

sorting1 :: Sort a -> Sort (f a)

default sorting1 :: Deciding1 Sorting f => Sort a -> Sort (f a)

desc :: Sort a -> Sort a

sort :: Sorting a => [a] -> [a]

O(n). Sort a list using discrimination.

sort = sortWith id

sortWith :: Sorting b => (a -> b) -> [a] -> [a]

O(n). Sort a list with a Schwartzian transformation by using discrimination.

This linear time replacement for sortWith and sortOn uses discrimination.

sortingBag :: Foldable f => Sort k -> Sort (f k)

Construct a stable ordered discriminator that sorts a list as multisets of elements from another stable ordered discriminator.

The resulting discriminator only cares about the set of keys and their multiplicity, and is sorted as if we'd sorted each key in turn before comparing.

sortingSet :: Foldable f => Sort k -> Sort (f k)

Construct a stable ordered discriminator that sorts a list as sets of elements from another stable ordered discriminator.

The resulting discriminator only cares about the set of keys, and is sorted as if we'd sorted each key in turn before comparing.

sortingCompare :: Sorting a => a -> a -> Ordering

Valid definition for compare in terms of Sorting.

Container Construction

toMap :: Sorting k => [(k, v)] -> Map k v

O(n). Construct a Map.

This is an asymptotically faster version of fromList, which exploits ordered discrimination.

>>> toMap []
fromList []

>>> toMap [(5,"a"), (3 :: Int,"b"), (5, "c")]
fromList [(3,"b"),(5,"c")]

>>> Map.fromList [(5,"a"), (3 :: Int,"b"), (5, "c")]
fromList [(3,"b"),(5,"c")]

>>> toMap [(5,"c"), (3,"b"), (5 :: Int, "a")]
fromList [(3,"b"),(5,"a")]

>>> Map.fromList [(5,"c"), (3,"b"), (5 :: Int, "a")]
fromList [(3,"b"),(5,"a")]

toMapWith :: Sorting k => (v -> v -> v) -> [(k, v)] -> Map k v

O(n). Construct a Map, combining values.

This is an asymptotically faster version of fromListWith, which exploits ordered discrimination.

(Note: values combine in anti-stable order for compatibility with fromListWith)

>>> toMapWith (++) [(5,"a"), (5,"b"), (3,"b"), (3,"a"), (5 :: Int,"c")]
fromList [(3,"ab"),(5,"cba")]

>>> Map.fromListWith (++) [(5,"a"), (5,"b"), (3,"b"), (3,"a"), (5 :: Int,"c")]
fromList [(3,"ab"),(5,"cba")]

>>> toMapWith (++) []
fromList []

toMapWithKey :: Sorting k => (k -> v -> v -> v) -> [(k, v)] -> Map k v

O(n). Construct a Map, combining values with access to the key.

This is an asymptotically faster version of fromListWithKey, which exploits ordered discrimination.

(Note: the values combine in anti-stable order for compatibility with fromListWithKey)

>>> let f key new_value old_value = show key ++ ":" ++ new_value ++ "|" ++ old_value
>>> toMapWithKey f [(5,"a"), (5,"b"), (3,"b"), (3,"a"), (5 :: Int,"c")]
fromList [(3,"3:a|b"),(5,"5:c|5:b|a")]

>>> toMapWithKey f []
fromList []

toIntMap :: [(Int, v)] -> IntMap v

O(n). Construct an IntMap.

>>> toIntMap []
fromList []

>>> toIntMap [(5,"a"), (3,"b"), (5, "c")]
fromList [(3,"b"),(5,"c")]

>>> IntMap.fromList [(5,"a"), (3,"b"), (5, "c")]
fromList [(3,"b"),(5,"c")]

>>> toIntMap [(5,"c"), (3,"b"), (5, "a")]
fromList [(3,"b"),(5,"a")]

>>> IntMap.fromList [(5,"c"), (3,"b"), (5, "a")]
fromList [(3,"b"),(5,"a")]

toIntMapWith :: (v -> v -> v) -> [(Int, v)] -> IntMap v

O(n). Construct an IntMap, combining values.

This is an asymptotically faster version of fromListWith, which exploits ordered discrimination.

(Note: values combine in anti-stable order for compatibility with fromListWith)

>>> toIntMapWith (++) [(5,"a"), (5,"b"), (3,"b"), (3,"a"), (5,"c")]
fromList [(3,"ab"),(5,"cba")]

>>> IntMap.fromListWith (++) [(5,"a"), (5,"b"), (3,"b"), (3,"a"), (5,"c")]
fromList [(3,"ab"),(5,"cba")]

>>> toIntMapWith (++) []
fromList []

toIntMapWithKey :: (Int -> v -> v -> v) -> [(Int, v)] -> IntMap v

O(n). Construct a Map, combining values with access to the key.

This is an asymptotically faster version of fromListWithKey, which exploits ordered discrimination.

(Note: the values combine in anti-stable order for compatibility with fromListWithKey)

>>> let f key new_value old_value = show key ++ ":" ++ new_value ++ "|" ++ old_value
>>> toIntMapWithKey f [(5,"a"), (5,"b"), (3,"b"), (3,"a"), (5,"c")]
fromList [(3,"3:a|b"),(5,"5:c|5:b|a")]

>>> IntMap.fromListWithKey f [(5,"a"), (5,"b"), (3,"b"), (3,"a"), (5,"c")]
fromList [(3,"3:a|b"),(5,"5:c|5:b|a")]

>>> toIntMapWithKey f []
fromList []

toSet :: Sorting k => [k] -> Set k

O(n). Construct a Set in linear time.

This is an asymptotically faster version of fromList, which exploits ordered discrimination.

toIntSet :: [Int] -> IntSet

O(n). Construct an IntSet in linear time.

This is an asymptotically faster version of fromList, which exploits ordered discrimination.

Joins

joining

Arguments

:: Discriminating f

=> f d
the discriminator to use

-> ([a] -> [b] -> c)
how to join two tables

-> (a -> d)
selector for the left table

-> (b -> d)
selector for the right table

-> [a]
left table

-> [b]
right table

-> [c]

O(n). Perform a full outer join while explicit merging of the two result tables a table at a time.

The results are grouped by the discriminator.

inner

Arguments

:: Discriminating f

=> f d
the discriminator to use

-> (a -> b -> c)
how to join two rows

-> (a -> d)
selector for the left table

-> (b -> d)
selector for the right table

-> [a]
left table

-> [b]
right table

-> [[c]]

O(n). Perform an inner join, with operations defined one row at a time.

The results are grouped by the discriminator.

This takes operation time linear in both the input and result sets.

outer

Arguments

:: Discriminating f

=> f d
the discriminator to use

-> (a -> b -> c)
how to join two rows

-> (a -> c)
row present on the left, missing on the right

-> (b -> c)
row present on the right, missing on the left

-> (a -> d)
selector for the left table

-> (b -> d)
selector for the right table

-> [a]
left table

-> [b]
right table

-> [[c]]

O(n). Perform a full outer join with operations defined one row at a time.

The results are grouped by the discriminator.

This takes operation time linear in both the input and result sets.

leftOuter

Arguments

:: Discriminating f

=> f d
the discriminator to use

-> (a -> b -> c)
how to join two rows

-> (a -> c)
row present on the left, missing on the right

-> (a -> d)
selector for the left table

-> (b -> d)
selector for the right table

-> [a]
left table

-> [b]
right table

-> [[c]]

O(n). Perform a left outer join with operations defined one row at a time.

The results are grouped by the discriminator.

This takes operation time linear in both the input and result sets.

rightOuter

Arguments

:: Discriminating f

=> f d
the discriminator to use

-> (a -> b -> c)
how to join two rows

-> (b -> c)
row present on the right, missing on the left

-> (a -> d)
selector for the left table

-> (b -> d)
selector for the right table

-> [a]
left table

-> [b]
right table

-> [[c]]

O(n). Perform a right outer join with operations defined one row at a time.

The results are grouped by the discriminator.

This takes operation time linear in both the input and result sets.