Data.Discrimination.Type

Documentation

newtype Disc a

Discriminator

TODO: use [(a,b)] -> [NonEmpty b] to better indicate safety?

Constructors

Disc

Fields

(%) :: forall b. [(a, b)] -> [[b]]

discNat :: Int -> Disc Int

unordered discrimination by bucket, not inherently thread-safe but it only has to restore the elements that were used unlike the more obvious ST implementation, so it wins by a huge margin in a race, especially when we have a large keyspace, sparsely used, with low contention. This will leak a number of arrays equal to the maximum concurrent contention for this resource. If this becomes a bottleneck we can make multiple stacks of working pads and index the stack with the hash of the current thread id to reduce contention at the expense of taking more memory.

bdiscNat :: Int -> ([v] -> v -> [v]) -> [(Int, v)] -> [[v]]